requirements.txt
numpy==1.8.0
nose==1.3.0
lettuce==0.2.19
simpy==3.0.5
redis==2.9.1
flask==0.10.1
flask-socketio==0.3.5
zato-redis-paginator==1.0.0
celery==3.1.11